Move internal/dict -> config/dict
