+++ cmd/internal/register/caches_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/go-spatial/tegola/cmd/internal/register"
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 )
 
 func TestCaches(t *testing.T) {
