+++ config/dict/types.go
