+++ provider/debug/debug.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/go-spatial/geom"
 	"github.com/go-spatial/tegola"
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 	"github.com/go-spatial/tegola/provider"
 )
