+++ provider/gpkg/cgo_test.go
@@ -5,7 +5,7 @@ package gpkg
 import (
 	"testing"
 
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 	"github.com/go-spatial/tegola/provider"
 )
