+++ provider/gpkg/gpkg_register_internal_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/go-spatial/geom"
 	"github.com/go-spatial/geom/cmp"
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 
 	_ "github.com/mattn/go-sqlite3"
 )
