+++ provider/postgis/postgis.go
@@ -16,7 +16,7 @@ import (
 	"github.com/go-spatial/tegola"
 	"github.com/go-spatial/tegola/provider"
 
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 )
 
 const Name = "postgis"
