+++ provider/provider.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/go-spatial/geom"
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 	"github.com/go-spatial/tegola/internal/log"
 )
