+++ provider/test/provider.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-spatial/tegola"
 	"github.com/go-spatial/tegola/provider"
 
-	"github.com/go-spatial/tegola/internal/dict"
+	"github.com/go-spatial/tegola/config/dict"
 )
 
 const Name = "test"
